+++ tests/geometry/volumes/testDigitalConvexity.cpp
@@ -1024,3 +1024,167 @@ SCENARIO( "DigitalConvexity< Z3 > full subconvexity of points and triangles", "[
   }
   
 }
+
+SCENARIO( "DigitalConvexity< Z3 > full covering of segments", "[full_cover][3d]" )
+{
+  typedef KhalimskySpaceND<3,int>          KSpace;
+  typedef KSpace::Point                    Point;
+  typedef KSpace::Vector                   Vector;
+  typedef KSpace::Space                    Space;
+  typedef HyperRectDomain< Space >         Domain;
+  typedef DigitalConvexity< KSpace >       DConvexity;
+
+  Domain     domain( Point( -20, -20, -20 ), Point( 20, 20, 20 ) );
+  DConvexity dconv ( Point( -21, -21, -21 ), Point( 21, 21, 21 ) );
+  {
+    Point a( 0, 0, 0 );
+    Point b( 3, 2, 1 );
+    auto LS = dconv.CoverCvxH( a, b );
+    auto P  = LS.toPointRange();
+    CAPTURE( P );
+    REQUIRE( P.size() == 9 );
+  }
+  {
+    Point a( 0, 0, 3 );
+    Point b( 3, 0, 0 );
+    auto LS = dconv.CoverCvxH( a, b );
+    auto P  = LS.toPointRange();
+    CAPTURE( P );
+    REQUIRE( P.size() == 7 );
+  }
+  {
+    Point a( 3, 0, 0 );
+    Point b( 0, 2, 5 );
+    auto LS = dconv.CoverCvxH( a, b );
+    auto P  = LS.toPointRange();
+    CAPTURE( P );
+    REQUIRE( P.size() == 17 );
+  }
+}
+
+SCENARIO( "DigitalConvexity< Z3 > full covering of triangles", "[full_cover][3d]" )
+{
+  typedef KhalimskySpaceND<3,int>          KSpace;
+  typedef KSpace::Point                    Point;
+  typedef KSpace::Vector                   Vector;
+  typedef KSpace::Space                    Space;
+  typedef HyperRectDomain< Space >         Domain;
+  typedef DigitalConvexity< KSpace >       DConvexity;
+
+  Domain     domain( Point( -20, -20, -20 ), Point( 20, 20, 20 ) );
+  DConvexity dconv ( Point( -21, -21, -21 ), Point( 21, 21, 21 ) );
+  {
+    Point a( 0, 0, 0 );
+    Point b( 2, 1, 0 );
+    Point c( 2, 2, 0 );  
+    auto LS = dconv.CoverCvxH( a, b, c );
+    auto P  = LS.toPointRange();
+    CAPTURE( P );
+    REQUIRE( P.size() == 10 );
+  }
+  {
+    Point a( 0, 0, 1 );
+    Point b( 2, 1, 0 );
+    Point c( 2, 2, 0 );  
+    auto LS = dconv.CoverCvxH( a, b, c );
+    auto P  = LS.toPointRange();
+    CAPTURE( P );
+    REQUIRE( P.size() == 10 );
+  }
+  {
+    Point a( 1, 0, 0 );
+    Point b( 0, 1, 0 );
+    Point c( 0, 0, 1 );  
+    auto LS = dconv.CoverCvxH( a, b, c );
+    auto P  = LS.toPointRange();
+    CAPTURE( P );
+    REQUIRE( P.size() == 7 );
+  }
+}
+
+
+
+SCENARIO( "DigitalConvexity< Z3 > full subconvexity and full covering of triangles", "[subconvexity][3d][full_cover]" )
+{
+  typedef KhalimskySpaceND<3,int>          KSpace;
+  typedef KSpace::Point                    Point;
+  typedef KSpace::Vector                   Vector;
+  typedef KSpace::Space                    Space;
+  typedef HyperRectDomain< Space >         Domain;
+  typedef DigitalConvexity< KSpace >       DConvexity;
+
+  Domain     domain( Point( -20, -20, -20 ), Point( 20, 20, 20 ) );
+  DConvexity dconv ( Point( -21, -21, -21 ), Point( 21, 21, 21 ) );
+
+  WHEN( "Computing many tetrahedra" ) {
+    const unsigned int nb     = 50;
+    unsigned int nb_total     = 0;
+    unsigned int nb_ok_tri    = 0;
+    unsigned int nb_ok_seg    = 0;
+    unsigned int nb_ko_seg    = 0;
+    unsigned int nb_subconvex = 0;
+    unsigned int nb_covered   = 0;
+    for ( unsigned int l = 0; l < nb; ++l )
+      {
+        const Point a { (rand() % 10 - 10), (rand() % 10 - 10), (rand() % 10 - 10) };
+        const Point b { (rand() % 20     ), (rand() % 20 - 10), (rand() % 20 - 10) };
+        const Point c { (rand() % 20 - 10), (rand() % 20     ), (rand() % 20 - 10) };
+        const Point d { (rand() % 20 - 10), (rand() % 20 - 10), (rand() % 20     ) };
+        const std::vector<Point> pts = { a, b, c, d };
+        const bool fulldim = dconv.isSimplexFullDimensional(pts.cbegin(), pts.cend());
+        if ( ! fulldim ) continue;
+        auto simplex = dconv.makeSimplex( pts.cbegin(), pts.cend() );
+        auto cover   = dconv.makeCellCover( simplex, 0, 3 );
+	auto ls      = dconv.StarCvxH( pts );
+        {
+          for ( unsigned int i = 0; i < 100; i++ )
+	    {
+	      const Point p { (rand() % 10 - 5), (rand() % 10 - 5), (rand() % 10 - 5) };
+	      const Point q { (rand() % 10 - 5), (rand() % 10 - 5), (rand() % 10 - 5) };
+	      const Point r { (rand() % 10 - 5), (rand() % 10 - 5), (rand() % 10 - 5) };
+	      const Vector n = ( q - p ).crossProduct( r - p );
+	      if ( n == Vector::zero ) continue;
+	      auto tri1 = dconv.makeSimplex( { p, q, r } );
+	      bool ok1  = dconv.isFullySubconvex( p, q, r, ls );
+	      bool ok2  = dconv.isFullyCovered( p, q, r, ls );
+	      if ( ok2 )
+		{
+		  // check covering of segments
+		  if ( dconv.isFullyCovered( p, q, ls ) ) nb_ok_seg += 1;
+		  else nb_ko_seg += 1;
+		  if ( dconv.isFullyCovered( p, r, ls ) ) nb_ok_seg += 1;
+		  else nb_ko_seg += 1;
+		  if ( dconv.isFullyCovered( q, r, ls ) ) nb_ok_seg += 1;
+		  else nb_ko_seg += 1;
+		}
+	      
+	      nb_subconvex += ok1 ? 1 : 0;
+	      nb_covered   += ok2 ? 1 : 0;		  
+	      bool ok = ok1 == ok2;
+	      if ( ! ok )
+		{
+		  std::cout << "***** FULL SUBCONVEXITY ERROR ON TRIANGLE ****" << std::endl;
+		  std::cout << "splx v =" << a << b << c << d << std::endl;
+		  std::cout << "simplex=" << simplex << std::endl;
+		  std::cout << "tri v  =" << p << q << r << std::endl;
+		  std::cout << "tri1=" << tri1 << std::endl;
+		  std::cout << "3 points are fully subconvex: " << ( ok1 ? "YES" : "NO" ) << std::endl;
+		  std::cout << "3 points are fully covered by star: " << ( ok2 ? "YES" : "NO" ) << std::endl;
+		}
+	      nb_ok_tri += ( ok ) ? 1 : 0;
+	      nb_total  += 1;
+	    }
+	}
+      }
+    THEN( "The number of subconvex and covered triangles should be equal." ) {
+      REQUIRE( nb_subconvex == nb_covered );
+    }
+    THEN( "Full subconvexity and covering should agree on every subset." ) {
+      REQUIRE( nb_ok_tri == nb_total );
+    }
+    THEN( "When a triangle is fully covered, its edges are fully covered also." ) {
+      REQUIRE( nb_ko_seg == 0 );
+    }
+  }
+  
+}